Move osimCamera into module for access during model conversion and for now create a camera attached to every segment in model
