+++ osimConverters/__init__.py
@@ -4,6 +4,7 @@ from .convertC3D2Gltf import *
 from .convertMotForce2Gltf import *
 from .convertOsim2Gltf import *
 from .convertOsimZip2Gltf import *
+from .osimCamera import *
 
 def convertNativeFileToGLTF(inFilePathOrURL):
         filename, file_extension = os.path.splitext(inFilePathOrURL)
